services/rag-service/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crawl4ai==0.4.247
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
fake-http-header==0.3.5
fastapi==0.115.6
filelock==3.17.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib-metadata==8.6.1
iniconfig==2.0.0
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.1
langchain-pinecone==0.2.2
langchain-tests==0.3.9
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.3.1
litellm==1.53.3
lxml==5.3.0
markupsafe==3.0.2
marshmallow==3.25.1
mockito==1.5.3
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
openai==1.59.9
orjson==3.10.15
packaging==24.2
pdf2image==1.16.3
pillow==10.4.0
pinecone==5.4.2
pinecone-client==5.0.1
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pip==23.2.1
playwright==1.49.1
pluggy==1.5.0
propcache==0.2.1
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
pyee==12.0.0
pyopenssl==25.0.0
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mockito==0.0.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.6
pyyaml==6.0.2
rank-bm25==0.2.2
rapidfuzz==3.11.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
sqlalchemy==2.0.37
starlette==0.41.3
syrupy==4.8.1
tabulate==0.9.0
tenacity==9.0.0
tf-playwright-stealth==1.1.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
unstructured==0.10.30
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
